% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{mxReport}
\alias{mxReport}
\title{Generate an OpenMx Report}
\usage{
mxReport(.models, report_file = "openmx_report.html", report_dir = ".",
  params = report_params(name = "Aaron"))
}
\arguments{
\item{models}{a list of `MxModel` objects}
}
\description{
Generate an OpenMx Report
}
\examples{
data(AdeFit)
mxReport(.models = list(AdeFit, AeFit))
}
