% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-module.r
\name{reportModule}
\alias{reportModule}
\alias{reportModule.MxModel}
\alias{reportModule.list}
\alias{reportModule.default}
\title{Generate appropriate report module}
\usage{
reportModule(x)

\method{reportModule}{MxModel}(x)

\method{reportModule}{list}(x)

\method{reportModule}{default}(x)
}
\description{
Generate appropriate report module
}
