% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.r
\name{smData}
\alias{smData}
\title{Extract submodel data}
\usage{
smData(x, rm.invar = TRUE, num.only = TRUE)
}
\arguments{
\item{x}{MxModel object}

\item{rm.invar}{remove variables with invariant values}

\item{num.only}{remove non-numeric variables}
}
\description{
Extract submodel data
}
